requirements.txt
google-api-python-client
requests~=2.28.1
bs4~=0.0.1
ics~=0.7.2
python-telegram-bot
selenium~=4.5.0
praw~=7.6.0
gTTS==2.2.1
# Alpha version fixes googletrans bug - see https://stackoverflow.com/a/65109346/14282680
googletrans==3.1.0a0
schedule~=1.1.0
feedparser~=6.0.10
lxml
PyMySQL~=1.0.2
python-dotenv~=0.21.0
yfinance>=0.1.63
mplfinance~=0.12.9b1
dateparser~=1.1.1
pony~=0.7.16
beautifulsoup4~=4.11.1
pytz~=2022.4